SS-9007: Improve icons rendering logic
