+++ types/store/shapediverStoreViewportIcons.ts
@@ -1,15 +1,4 @@
-import ArButton from "@AppBuilderShared/components/shapediver/viewport/buttons/ArButton";
-import CamerasButton from "@AppBuilderShared/components/shapediver/viewport/buttons/CamerasButton";
-import FullscreenButton from "@AppBuilderShared/components/shapediver/viewport/buttons/FullscreenButton";
-import HistoryMenuButton from "@AppBuilderShared/components/shapediver/viewport/buttons/HistoryMenuButton";
-import RedoButton from "@AppBuilderShared/components/shapediver/viewport/buttons/RedoButton";
-import ReloadButton from "@AppBuilderShared/components/shapediver/viewport/buttons/ReloadButton";
-import {CommonButtonProps} from "@AppBuilderShared/components/shapediver/viewport/buttons/types";
-import UndoButton from "@AppBuilderShared/components/shapediver/viewport/buttons/UndoButton";
-import ZoomButton from "@AppBuilderShared/components/shapediver/viewport/buttons/ZoomButton";
-import {DividerProps} from "@mantine/core";
 import {IViewportApi} from "@shapediver/viewer.viewport";
-import {ReactNode} from "react";
 
 export enum ViewportIconButtonEnum {
 	Ar = "ar",
@@ -67,30 +56,4 @@ export interface IShapeDiverViewportIconsStore {
 
 	// clear all icons for a viewport
 	clear: (viewportId: string) => void;
-	render: (
-		viewportId: string,
-		buttonContext: ButtonRenderContext,
-		dividerProps: DividerProps,
-	) => ReactNode[];
 }
-
-export interface ButtonRenderContext extends CommonButtonProps {
-	viewport?: any;
-	namespace?: string;
-	buttonsDisabled: boolean;
-	executing: boolean;
-	hasPendingChanges: boolean;
-	iconsVisible: boolean;
-	fullscreenId: string;
-}
-
-export const ViewportTypeToIcon = {
-	[ViewportIconButtonEnum.Ar]: ArButton,
-	[ViewportIconButtonEnum.Zoom]: ZoomButton,
-	[ViewportIconButtonEnum.Fullscreen]: FullscreenButton,
-	[ViewportIconButtonEnum.Cameras]: CamerasButton,
-	[ViewportIconButtonEnum.Undo]: UndoButton,
-	[ViewportIconButtonEnum.Redo]: RedoButton,
-	[ViewportIconButtonEnum.Reload]: ReloadButton,
-	[ViewportIconButtonEnum.HistoryMenu]: HistoryMenuButton,
-};
